src refactor/add utils, add run dev without docker

+++ local-setup-and-run.sh
@@ -18,6 +18,9 @@ curl -X POST https://api.twilio.com/2010-04-01/Accounts/$TWILIO_ACCOUNT_SID/Inco
 
 echo 
 
-docker build --tag=whatsapp .
-
-docker run --rm -it -p 3000:3000 --env-file=.env -v ~/.aws:/root/.aws whatsapp
+if [ "$1" == "docker" ]; then
+    docker build --tag=whatsapp .
+    docker run --rm -it -p 3000:3000 --env-file=.env -v ~/.aws:/root/.aws whatsapp
+else
+    ts-node -r dotenv/config --project ./tsconfig.json --files src/run.ts
+fi
